移除了矩形类中不安全的联合体 提高圆弧包围盒的精度 性能提高了一半

+++ nvcv2/template/template.cpp
@@ -54,10 +54,10 @@ real_t template_match(const Image<Binary> & src, const Image<Binary> & tmp, cons
     // uint base = tmp.sum() / 255;
     // uint or_score = 0;
 
-    for(uint y = 0; y < (uint)rect.h; y++){
+    for(uint y = 0; y < rect.h(); y++){
         const auto * tmp_ptr = &tmp[Vector2i(0,y)];
         const auto * src_ptr = &src[Vector2i(0,y) + offs];
-        for(uint x = 0; x < (uint)rect.w; x++){
+        for(uint x = 0; x < rect.w(); x++){
             score += int(bool(*tmp_ptr) ^ bool(*src_ptr));
             // if(bool(*tmp_ptr) + bool(*src_ptr) == 1) score++;
             tmp_ptr++;
